Sweeping changes
* Add /dist/ to .gitignore (remove later)
* Modify webpack.config.js file (uncomment later)
* Start to modularize ag_grid.js code
