+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 yarn-error.log
 .eslintrc.js
 .eslintrc.json
+/dist/
