+++ src/examples/ag_grid/agColumn.js
@@ -1,0 +1,31 @@
+class AgColumn {
+  constructor(config) {
+    this.config = config;
+    this.formatColumns();
+  }
+
+  // Format the columns based on the queryResponse into an object ag-grid can handle.
+  formatColumns() {
+    const { queryResponse } = globalConfig;
+    const { pivots, measures, dimensions: dims } = queryResponse.fields;
+    const dimensions = basicDimensions(dims, this.config);
+
+    const tableCalcs = queryResponse.fields.table_calculations;
+
+    // Measures and table calcs are only shown in the context of pivots when present.
+    if (!_.isEmpty(pivots)) {
+      addPivots(dimensions, this.config);
+    } else {
+      // When there are no pivots, show measures and table calcs in own column.
+      if (!_.isEmpty(measures)) {
+        addMeasures(dimensions, measures, this.config);
+      }
+      if (!_.isEmpty(tableCalcs)) {
+        addTableCalculations(dimensions, tableCalcs);
+      }
+    }
+    this.formattedColumns = dimensions;
+  }
+}
+
+export default AgColumn;
