+++ src/examples/ag_grid/agData.js
@@ -1,0 +1,30 @@
+class AgData {
+  constructor(data, formattedColumns) {
+    this.data = data;
+    this.formattedColumns = formattedColumns;
+    this.formatData();
+  }
+
+  formatData() {
+    this.formattedData = this.data.map(datum => {
+      const formattedDatum = {};
+
+      this.formattedColumns.forEach(col => {
+        const { children, colType, field: colField, lookup } = col;
+        if (colType === 'row') { return; }
+
+        if (colType === 'pivot') {
+          children.forEach(child => {
+            formattedDatum[child.field] = displayData(datum[child.measure][child.pivotKey]);
+          });
+        } else {
+          formattedDatum[colField] = displayData(datum[lookup]);
+        }
+      });
+
+      return formattedDatum;
+    });
+  }
+}
+
+export default AgData;
