+++ src/examples/ag_grid/globalConfig.js
@@ -1,0 +1,19 @@
+class GlobalConfig {
+  constructor() {
+    this.selectedFields = [];
+  }
+
+  addSelectedField(field) {
+    if (this.selectedFields.indexOf(field) === -1) {
+      this.selectedFields.push(field);
+    }
+  }
+
+  removeSelectedField(field) {
+    this.selectedFields = this.selectedFields.filter(selectedField => {
+      return selectedField !== field;
+    });
+  }
+}
+
+export default GlobalConfig;
