+++ src/examples/ag_grid/options.js
@@ -1,0 +1,181 @@
+// These are all of the options that set up the initial 
+export const options = {
+  // FORMATTING
+  enableConditionalFormatting: {
+    default: false,
+    label: 'Enable Conditional Formatting',
+    order: 1,
+    section: 'Formatting',
+    type: 'boolean',
+  },
+  perColumnRange: {
+    default: true,
+    hidden: true,
+    label: 'Per column range',
+    order: 2,
+    section: 'Formatting',
+    type: 'boolean',
+  },
+  conditionalFormattingType: {
+    default: 'all',
+    display: 'select',
+    label: 'Formatting Type',
+    order: 3,
+    section: 'Formatting',
+    type: 'string',
+    values: [
+      { 'All': 'all' },
+      { 'Subtotals only': 'subtotals_only' },
+      { 'Non-subtotals only': 'non_subtotals_only' },
+    ],
+  },
+  includeNullValuesAsZero: {
+    default: false,
+    label: 'Include Null Values as Zero',
+    order: 4,
+    section: 'Formatting',
+    type: 'boolean',
+  },
+  formattingStyle: {
+    default: 'low_to_high',
+    display: 'select',
+    label: 'Format',
+    order: 5,
+    section: 'Formatting',
+    type: 'string',
+    values: [
+      { 'From low to high': 'low_to_high' },
+      { 'From high to low': 'high_to_low' },
+    ],
+  },
+  formattingPalette: {
+    default: 'red_yellow_green',
+    display: 'select',
+    label: 'Palette',
+    order: 6,
+    section: 'Formatting',
+    type: 'string',
+    values: [
+      { 'Red to Yellow to Green': 'red_yellow_green' },
+      { 'Red to White to Green': 'red_white_green' },
+      { 'Red to White': 'red_white' },
+      { 'White to Green': 'white_green' },
+      { 'Custom...': 'custom' },
+    ],
+  },
+  lowColor: {
+    display: 'color',
+    display_size: 'third',
+    label: 'Low', // These values updated in updateAsync
+    order: 7,
+    section: 'Formatting',
+    type: 'string',
+  },
+  midColor: {
+    display: 'color',
+    display_size: 'third',
+    label: 'Middle',
+    order: 8,
+    section: 'Formatting',
+    type: 'string',
+  },
+  highColor: {
+    display: 'color',
+    display_size: 'third',
+    label: 'High',
+    order: 9,
+    section: 'Formatting',
+    type: 'string',
+  },
+  applyTo: {
+    default: 'all_numeric_fields',
+    display: 'select',
+    label: 'Apply to',
+    order: 10,
+    section: 'Formatting',
+    type: 'string',
+    values: [
+      { 'All numeric fields': 'all_numeric_fields' },
+      { 'Select fields...': 'select_fields' },
+    ],
+  },
+  // CONFIG
+  fontSize: {
+    default: 12,
+    display_size: 'third',
+    label: 'Font size (pt)',
+    order: 1,
+    section: 'Config',
+    type: 'number',
+  },
+  fontFamily: {
+    default: 'Open Sans, Helvetica, Arial, sans-serif',
+    display: 'select',
+    display_size: 'two-thirds',
+    label: 'Font Family',
+    order: 2,
+    section: 'Config',
+    type: 'string',
+    values: [
+      { 'Looker': 'Open Sans, Helvetica, Arial, sans-serif' },
+      { 'Helvetica': 'BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif' },
+      { 'Times New Roman': 'Times, "Times New Roman", serif' },
+    ],
+  },
+  rowHeight: {
+    default: 25,
+    display_size: 'third',
+    label: 'Row Height',
+    order: 3,
+    section: 'Config',
+    type: 'number',
+  },
+  // SERIES
+  truncateColumnNames: {
+    default: false,
+    label: 'Truncate Column Names',
+    order: 1,
+    section: 'Series',
+    type: 'boolean',
+  },
+  showFullFieldName: {
+    default: false,
+    label: 'Show Full Field Name',
+    order: 2,
+    section: 'Series',
+    type: 'boolean',
+  },
+  // CUSTOMIZATIONS
+
+  // PLOT
+  theme: {
+    default: 'ag-theme-looker',
+    display: 'select',
+    label: 'Table Theme',
+    order: 1,
+    section: 'Plot',
+    type: 'string',
+    values: [
+      { Looker: 'ag-theme-looker' },
+      { Balham: 'ag-theme-balham' },
+      { Fresh: 'ag-theme-fresh' },
+      { Dark: 'ag-theme-dark' },
+      { Blue: 'ag-theme-blue' },
+      { Bootstrap: 'ag-theme-bootstrap' },
+    ],
+  },
+  showRowNumbers: {
+    default: false,
+    label: 'Show Row Numbers',
+    order: 2,
+    section: 'Plot',
+    type: 'boolean',
+  },
+  autoSizeEnabled: {
+    default: true,
+    label: 'Enable Auto Sizing',
+    order: 3,
+    section: 'Plot',
+    type: 'boolean',
+  },
+};
